+++ resources/views/questions/index.blade.php
@@ -41,10 +41,14 @@
                                             <a href="{{ $question->url }}">{{ $question->title }}</a>
                                         </h3>
                                         <div class="ml-auto">
-                                            <a class="btn btn-outline-info btn-sm" href="{!! route('questions.edit', $question->id) !!}">Edit</a>
-                                            {!! Form::open(['route' => ['questions.destroy', $question->id], 'method' => 'DELETE', 'class' => 'form-delete']) !!}
-                                            {!! Form::submit('delete',['class' => 'btn btn-sm btn-outline-danger', 'onclick' => 'return confirm("Are you sure?")']) !!}
-                                            {!! Form::close() !!}
+                                            @if(\Auth::user() && \Auth::user()->can('update-question', $question))
+                                                <a class="btn btn-outline-info btn-sm" href="{!! route('questions.edit', $question->id) !!}">Edit</a>
+                                            @endif
+                                            @if(\Auth::user() && \Auth::user()->can('delete-question', $question))
+                                                {!! Form::open(['route' => ['questions.destroy', $question->id], 'method' => 'DELETE', 'class' => 'form-delete']) !!}
+                                                {!! Form::submit('delete',['class' => 'btn btn-sm btn-outline-danger', 'onclick' => 'return confirm("Are you sure?")']) !!}
+                                                {!! Form::close() !!}
+                                            @endif
                                         </div>
                                     </div>
                                     <p class="lead">
